+++ modules/lambdas/main.tf
@@ -14,15 +14,10 @@ module "lambda_security_group" {
   ## Cumulus Variables
   ## --------------------------
   ## REQUIRED
-  prefix      = var.prefix
-  vpc_id      = var.vpc_id
+  prefix = var.prefix
+  vpc_id = var.vpc_id
   ## OPTIONAL
-  tags   = local.tags
-  ## --------------------------
-  ## ORCA Variables
-  ## --------------------------
-  ## OPTIONAL
-  database_port = var.database_port
+  tags = local.tags
 }
 
 # restore_object_arn - IAM module reference
@@ -37,7 +32,7 @@ module "restore_object_arn" {
   permissions_boundary_arn = var.permissions_boundary_arn
   prefix                   = var.prefix
   # OPTIONAL
-  tags   = local.tags
+  tags = local.tags
   # --------------------------
   # ORCA Variables
   # --------------------------
@@ -163,7 +158,7 @@ resource "aws_lambda_function" "request_files" {
       RESTORE_EXPIRE_DAYS      = var.orca_recovery_expiration_days
       RESTORE_REQUEST_RETRIES  = var.orca_recovery_retry_limit
       RESTORE_RETRY_SLEEP_SECS = var.orca_recovery_retry_interval
-      RESTORE_RETRIEVAL_TYPE   = var.orca_recovery_retrieval_type
+      RESTORE_RETRIEVAL_TYPE   = "Standard"
       DB_QUEUE_URL             = var.orca_sqs_status_update_queue_id
       ORCA_DEFAULT_BUCKET      = var.orca_default_bucket
     }
@@ -245,10 +240,7 @@ resource "aws_lambda_function" "post_to_database" {
 
   environment {
     variables = {
-      PREFIX           = var.prefix
-      DATABASE_PORT    = var.database_port
-      DATABASE_NAME    = var.database_name
-      APPLICATION_USER = var.database_app_user
+      PREFIX = var.prefix
     }
   }
 }
@@ -296,10 +288,7 @@ resource "aws_lambda_function" "request_status_for_granule" {
 
   environment {
     variables = {
-      PREFIX        = var.prefix
-      DATABASE_PORT = var.database_port
-      DATABASE_NAME = var.database_name
-      DATABASE_USER = var.database_app_user
+      PREFIX = var.prefix
     }
   }
 }
@@ -329,10 +318,7 @@ resource "aws_lambda_function" "request_status_for_job" {
 
   environment {
     variables = {
-      PREFIX        = var.prefix
-      DATABASE_PORT = var.database_port
-      DATABASE_NAME = var.database_name
-      DATABASE_USER = var.database_app_user
+      PREFIX = var.prefix
     }
   }
 }
@@ -359,9 +345,6 @@ resource "aws_lambda_function" "post_copy_request_to_queue" {
   environment {
     variables = {
       PREFIX             = var.prefix
-      DATABASE_PORT      = var.database_port
-      DATABASE_NAME      = var.database_name
-      APPLICATION_USER   = var.database_app_user
       DB_QUEUE_URL       = var.orca_sqs_status_update_queue_id
       RECOVERY_QUEUE_URL = var.orca_sqs_staged_recovery_queue_id
       MAX_RETRIES        = var.orca_recovery_retry_limit
@@ -451,12 +434,16 @@ resource "aws_lambda_function" "db_deploy" {
 
   environment {
     variables = {
-      PREFIX           = var.prefix
-      DATABASE_PORT    = var.database_port
-      DATABASE_NAME    = var.database_name
-      APPLICATION_USER = var.database_app_user
-      ADMIN_USER       = "postgres"
-      ADMIN_DATABASE   = "postgres"
+      PREFIX = var.prefix
     }
   }
+}
+
+## =============================================================================
+## NULL RESOURCES - 1x Use
+## =============================================================================
+data "aws_lambda_invocation" "db_migration" {
+  # depends_on = [aws_db_instance.postgresql]
+  function_name = aws_lambda_function.db_deploy.function_name
+  input = jsonencode({})
 }