+++ tasks/shared_libraries/database/shared_db.py
@@ -6,14 +6,23 @@ Description: Shared library for database objects needed by the various libraries
 
 import os
 import boto3
+import json
+import time
+import random
+import functools
 from sqlalchemy import create_engine
 from sqlalchemy.engine import URL
+from sqlalchemy.exc import OperationalError
 from sqlalchemy.future import Engine
 from cumulus_logger import CumulusLogger
-from typing import Any, List, Dict, Optional, Union
+from typing import Any, Dict, TypeVar, Callable
 
 # instantiate CumulusLogger
 logger = CumulusLogger(name="orca")
+MAX_RETRIES = 3  # number of times to retry.
+BACKOFF_FACTOR = 2  # Value of the factor used to backoff
+INITIAL_BACKOFF_IN_SECONDS = 1  # Number of seconds to sleep the first time through.
+RT = TypeVar('RT')  # return type
 
 
 def get_configuration() -> Dict[str, str]:
@@ -25,16 +34,9 @@ def get_configuration() -> Dict[str, str]:
     Environment Variables:
         PREFIX (str): Deployment prefix used to pull the proper AWS secret.
         AWS_REGION (str): AWS reserved runtime variable used to set boto3 client region.
-        DATABASE_PORT (str): The database port. The standard is 5432
-        DATABASE_NAME (str): The name of the application database.
-        APPLICATION_USER (str): The name of the database application user.
-        ADMIN_USER (str): *OPTIONAL* The name of the database super user (postgres).
-        ADMIN_DATABASE (str): *OPTIONAL* The name of the admin database for the instance (postgres).
 
     Parameter Store:
-        <prefix>-drdb-user-pass (string): The password for the application user (APPLICATION_USER).
-        <prefix>-drdb-host (string): The database host.
-        <prefix>-drdb-admin-pass: The password for the admin user
+        <prefix>-orca-db-login-secret (string): The json string containing all the admin and user db login info.
     ```
 
     Args:
@@ -65,56 +67,21 @@ def get_configuration() -> Dict[str, str]:
         logger.critical(message)
         raise Exception(message)
 
-    # Create config dictionary
-    config = {}
-
-    # Create my config map that maps a key to the environment variable name
-    config_map = {
-        "database": "DATABASE_NAME",
-        "port": "DATABASE_PORT",
-        "app_user": "APPLICATION_USER",
-        "admin_user": "ADMIN_USER",
-        "admin_database": "ADMIN_DATABASE",
-    }
-
-    # Get the environment variables
-    for key, value in config_map.items():
-        logger.debug(f"Getting environment variable {value} value.")
-        env_value = os.getenv(value, None)
-
-        if env_value is None or len(env_value) == 0:
-            # Set defaults for optional env variables
-            if key in ["admin_user", "admin_database"]:
-                env_value = "postgres"
-                logger.info(f"Setting variable {value} value to {env_value}")
-            else:
-                message = f"Environment variable {value} is not set and is required"
-                logger.critical(message)
-                raise Exception(message)
-
-        config[key] = env_value
-
-    # Get the secret variables
     try:
         logger.debug("Creating secretsmanager resource.")
         secretsmanager = boto3.client("secretsmanager", region_name=aws_region)
 
-        logger.debug("Retrieving database application user password.")
-        app_user_pw = secretsmanager.get_secret_value(
-            SecretId=f"{prefix}-drdb-user-pass"
+        logger.debug(
+            "Retrieving db login info for both user and admin as a dictionary."
         )
-        config["app_user_password"] = app_user_pw["SecretString"]
-
-        logger.debug("Retrieving database root user password.")
-        admin_user_pw = secretsmanager.get_secret_value(
-            SecretId=f"{prefix}-drdb-admin-pass"
+        config = json.loads(
+            secretsmanager.get_secret_value(SecretId=f"{prefix}-orca-db-login-secret")[
+                "SecretString"
+            ]
         )
-        config["admin_user_password"] = admin_user_pw["SecretString"]
-
-        logger.debug("Retrieving database host information.")
-        db_host = secretsmanager.get_secret_value(SecretId=f"{prefix}-drdb-host")
-        config["host"] = db_host["SecretString"]
-
+        logger.debug(
+            "Successfully retrieved db login info for both user and admin as a dictionary."
+        )
     except Exception as e:
         logger.critical("Failed to retrieve secret.", exc_info=True)
         raise Exception("Failed to retrieve secret manager value.")
@@ -165,8 +132,8 @@ def get_admin_connection(config: Dict[str, str], database: str = None) -> Engine
         host=config["host"],
         port=config["port"],
         database=admin_database,
-        username=config["admin_user"],
-        password=config["admin_user_password"],
+        username=config["admin_username"],
+        password=config["admin_password"],
     )
 
     return connection
@@ -188,9 +155,65 @@ def get_user_connection(config: Dict[str, str]) -> Engine:
     connection = _create_connection(
         host=config["host"],
         port=config["port"],
-        database=config["database"],
-        username=config["app_user"],
-        password=config["app_user_password"],
+        database=config["user_database"],
+        username=config["user_username"],
+        password=config["user_password"],
     )
 
     return connection
+
+
+# Retry decorator for functions
+def retry_operational_error(max_retries: int = MAX_RETRIES, backoff_in_seconds: int = INITIAL_BACKOFF_IN_SECONDS,
+                            backoff_factor: int = BACKOFF_FACTOR) -> Callable[[Callable[[], RT]], Callable[[], RT]]:
+    """
+    Decorator takes arguments to adjust number of retries and backoff strategy.
+    Args:
+        max_retries (int): number of times to retry in case of failure.
+        backoff_in_seconds (int): Number of seconds to sleep the first time through.
+        backoff_factor (int): Value of the factor used for backoff.
+    """
+
+    def decorator_retry_operational_error(func: Callable[[], RT]) -> Callable[[], RT]:
+        """
+        Main Decorator that takes our function as an argument
+        """
+
+        @functools.wraps(func)  # Use built in for decorators
+        def wrapper_retry_operational_error(*args, **kwargs) -> RT:
+            """
+            Wrapper that performs our extra tasks on the function
+            """
+            # Initialize the retry loop
+            total_retries = 0
+
+            # Enter loop
+            while True:
+                # Try the function and catch the expected error
+                try:
+                    return func(*args, **kwargs)
+                except OperationalError:
+                    if total_retries == max_retries:
+                        # Log it and re-raise if we maxed our retries + initial attempt
+                        logger.error(
+                            "Encountered Errors {total_attempts} times. Reached max retry limit.",
+                            total_attempts=total_retries)
+                        raise
+                    else:
+                        # perform exponential delay
+                        backoff_time = (backoff_in_seconds * backoff_factor ** total_retries + random.uniform(0, 1))
+                        logger.error(
+                            "Encountered OperationalError on attempt {total_attempts}. "
+                            "Sleeping {backoff_time} seconds.",
+                            total_attempts=total_retries, backoff_time=backoff_time)
+                        time.sleep(backoff_time)
+                        total_retries += 1
+                except Exception as ex:
+                    logger.error("Encountered a non-retriable error: {ex}", ex=ex)
+                    raise ex
+
+        # Return our wrapper
+        return wrapper_retry_operational_error
+
+    # Return our decorator
+    return decorator_retry_operational_error